Use bolts from UniProKit. Convert to square base area. Reduce height
because stability is more important for precision than the
maximum height. The square base area now is greater than before.

Also remove "Assembly-" prefix as it is maintenance heavy and the
distinction between part and assembly is already encoded in what is a
group instance and what is mesh (in the source file following linked
data blocks).